requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Login>=0.6.3
Flask-WTF>=1.2.0
WTForms>=3.1.0
email-validator>=2.1.0
gunicorn>=21.2.0
pytest>=7.4.0
python-dotenv>=1.0.0